Another chronicle optimization

+++ spring/src/main/java/com/d4dl/permean/data/CellRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.web.bind.annotation.CrossOrigin;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -41,5 +40,5 @@ public interface CellRepository extends PagingAndSortingRepository <Cell, String
      * @return
      */
     //@EntityGraph(value = "Cell.vertices", type = EntityGraph.EntityGraphType.FETCH)
-    List<Cell> findByCenterLatitudeBetweenAndCenterLongitudeBetween(@Param("top") BigDecimal topLat, @Param("bottom") BigDecimal bottomLat, @Param("east") BigDecimal eastLng, @Param("west") BigDecimal westLng);
+    List<Cell> findByCenterLatitudeBetweenAndCenterLongitudeBetween(@Param("top") float topLat, @Param("bottom") float bottomLat, @Param("east") float eastLng, @Param("west") float westLng);
 }
